+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 1.2 (2014-02-??)
 ================
+Enhancements to the backoff policies in `atmos.retries`:
+
+ - Removed the previous backoff value from the list of items provided to a `BackoffPolicy` when calculating a backoff and replaced it with the most recently thrown exception.
+ 
 Enhancements to the termination policies in `atmos.retries`:
 
  - Added `TerminationPolicy.ImmediatelyTerminate`, a policy that never retries.